feat(cli): 将init抽出来sdk化
